requirements.txt
albumentations>=1.3.1
matplotlib>=3.5.0
numpy>=1.20.0
Pillow>=9.0.0
scikit-learn>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
wandb>=0.15.0
